rename pkg signer to security
